+++ be15_DevEagles_FE/src/features/schedules/components/ScheduleCalendar.vue
@@ -1,10 +1,12 @@
 <script setup>
-  import { defineProps, watch, ref } from 'vue';
+  import { defineProps, watch, ref, nextTick } from 'vue';
   import FullCalendar from '@fullcalendar/vue3';
   import dayGridPlugin from '@fullcalendar/daygrid';
   import timeGridPlugin from '@fullcalendar/timegrid';
   import interactionPlugin from '@fullcalendar/interaction';
   import koLocale from '@fullcalendar/core/locales/ko';
+  import tippy from 'tippy.js';
+  import 'tippy.js/dist/tippy.css';
 
   const props = defineProps({
     schedules: {
@@ -15,10 +17,22 @@
 
   const emit = defineEmits(['clickSchedule']);
 
+  const staffColors = {
+    최민수: 'var(--color-primary-main)',
+    이채은: '#60bafa',
+    김민지: '#f4e0ab',
+  };
+
+  const statusColors = {
+    '예약 확정': '#c6f6d5',
+    '예약 대기': '#fefcbf',
+  };
+
   const calendarOptions = ref({
     plugins: [dayGridPlugin, timeGridPlugin, interactionPlugin],
     initialView: 'dayGridMonth',
     editable: true,
+    eventDurationEditable: false,
     locale: koLocale,
     headerToolbar: {
       left: 'prev,next today',
@@ -26,26 +40,94 @@
       right: 'dayGridMonth,timeGridWeek,timeGridDay',
     },
     events: props.schedules,
-    eventContent({ event }) {
-      const dotColor = event.backgroundColor || '#999';
-      const title = event.title;
+
+    eventContent({ event, view }) {
+      const type = event.extendedProps.type;
+      const time = event.start.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
+      const viewType = view.type;
+
       return {
         html: `
-        <div style="display: flex; align-items: center; gap: 6px;">
-          <span style="display:inline-block;width:10px;height:10px;background:${dotColor};border-radius:2px;"></span>
-          <span style="color:#111;font-size:13px;font-weight:500;">${title}</span>
+        <div class="custom-event-style ${type}">
+          ${type === 'reservation' || type === 'holiday' ? '<div class="left-bar"></div>' : ''}
+          <div class="text">
+            ${viewType === 'dayGridMonth' ? `<strong>${time}</strong>&nbsp;` : ''}
+            ${event.title}
+          </div>
         </div>
       `,
       };
     },
+
     eventClick(info) {
       emit('clickSchedule', info.event.id);
     },
-    eventDrop(info) {
-      console.log('이벤트 이동됨:', {
-        id: info.event.id,
-        newStart: info.event.start,
-        newEnd: info.event.end,
+
+    eventDidMount(info) {
+      const { event, el, view } = info;
+      const { type, status, staff, customer, service, timeRange, memo } = event.extendedProps;
+
+      const viewType = view.type;
+      const borderColor = staffColors[staff?.trim()] || '#999';
+      const bgColor =
+        type === 'reservation'
+          ? statusColors[status] || '#e2e8f0'
+          : type === 'holiday'
+            ? 'rgba(255, 0, 0, 0.3)'
+            : '#fff';
+
+      const box = el.querySelector('.custom-event-style');
+      const bar = el.querySelector('.left-bar');
+
+      if (box) {
+        box.style.backgroundColor = bgColor;
+        if (type === 'event') {
+          box.style.border = `2px solid ${borderColor}`;
+        }
+      }
+
+      if (bar && (type === 'reservation' || type === 'holiday')) {
+        bar.style.backgroundColor = borderColor;
+        if (viewType === 'dayGridMonth') {
+          nextTick(() => {
+            bar.style.height = `${box.offsetHeight}px`;
+          });
+        }
+      }
+
+      // Tooltip 구성
+      const tooltipLines = [];
+
+      const fallbackTimeRange =
+        event.start && event.end
+          ? `${event.start.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })} - ${event.end.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}`
+          : null;
+
+      if (type === 'reservation') {
+        tooltipLines.push(
+          customer && `👤 고객: ${customer}`,
+          service && `💈 시술: ${service}`,
+          (timeRange || fallbackTimeRange) && `🕒 시간: ${timeRange || fallbackTimeRange}`,
+          memo && `📝 메모: ${memo}`
+        );
+      } else if (type === 'event') {
+        tooltipLines.push(
+          (timeRange || fallbackTimeRange) && `🕒 시간: ${timeRange || fallbackTimeRange}`,
+          memo && `📝 메모: ${memo}`
+        );
+      } else if (type === 'holiday') {
+        tooltipLines.push(`📅 종일 휴무`, memo && `📝 메모: ${memo}`);
+      }
+
+      const tooltip = tooltipLines.filter(Boolean).join('<br/>');
+
+      tippy(el, {
+        content: tooltip || '스케줄 정보 없음',
+        allowHTML: true,
+        placement: 'top',
+        animation: 'shift-away',
+        delay: [100, 0],
+        theme: 'light',
       });
     },
   });
@@ -53,7 +135,21 @@
   watch(
     () => props.schedules,
     newVal => {
-      calendarOptions.value.events = newVal;
+      calendarOptions.value.events = newVal.map(item => ({
+        id: item.id,
+        title: item.title,
+        start: item.start,
+        end: item.end,
+        extendedProps: {
+          type: item.type,
+          status: item.status,
+          staff: item.staff,
+          customer: item.customer,
+          service: item.service,
+          timeRange: item.timeRange,
+          memo: item.memo,
+        },
+      }));
     },
     { immediate: true, deep: true }
   );
@@ -65,8 +161,62 @@
 
 <style scoped>
   :deep(.fc-event) {
+    background-color: transparent !important;
+    border: none !important;
     font-weight: 500;
-    border: none;
     border-radius: 6px;
+    box-shadow: none;
+  }
+
+  :deep(.fc-daygrid-event) {
+    display: block !important;
+    width: 100% !important;
+    max-width: 100% !important;
+  }
+
+  :deep(.custom-event-style) {
+    position: relative;
+    display: flex;
+    align-items: center;
+    font-size: 13px;
+    font-weight: 500;
+    padding: 6px 8px;
+    border-radius: 6px;
+    overflow: hidden;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    height: 100%;
+  }
+
+  :deep(.custom-event-style.event) {
+    background-color: #fff;
+  }
+
+  :deep(.custom-event-style .left-bar) {
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 6px;
+    border-radius: 6px 0 0 6px;
+    height: 100%;
+  }
+
+  :deep(.custom-event-style .text) {
+    padding-left: 10px;
+    z-index: 1;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    color: #111;
+  }
+
+  /* Tippy Tooltip */
+  .tippy-box[data-theme~='light'] {
+    background-color: white;
+    color: #333;
+    font-size: 13px;
+    padding: 8px;
+    border-radius: 6px;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
   }
 </style>
