+++ be15_DevEagles_FE/src/features/schedules/views/ReservationHistoryView.vue
@@ -25,7 +25,7 @@
   const columns = [
     { key: 'name', title: '고객 이름', width: '120px' },
     { key: 'service', title: '시술', width: '100px' },
-    { key: 'employee', title: '담당자', width: '100px' },
+    { key: 'staff', title: '담당자', width: '100px' },
     { key: 'date', title: '예약 날짜', width: '160px' },
     { key: 'change', title: '변경 내용', width: '100px' },
     { key: 'changedAt', title: '변경 일시', width: '160px' },
@@ -36,7 +36,7 @@
       id: 1,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
@@ -45,7 +45,7 @@
       id: 2,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -54,7 +54,7 @@
       id: 3,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -63,7 +63,7 @@
       id: 4,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
@@ -72,7 +72,7 @@
       id: 5,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -81,7 +81,7 @@
       id: 6,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -90,7 +90,7 @@
       id: 7,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
@@ -99,7 +99,7 @@
       id: 8,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
@@ -108,7 +108,7 @@
       id: 9,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
@@ -117,7 +117,7 @@
       id: 10,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -126,7 +126,7 @@
       id: 11,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -135,7 +135,7 @@
       id: 12,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
@@ -144,7 +144,7 @@
       id: 13,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -153,7 +153,7 @@
       id: 14,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '수정',
       changedAt: '2025.06.08 11:08',
@@ -162,7 +162,7 @@
       id: 15,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
@@ -171,7 +171,7 @@
       id: 16,
       name: '김미글',
       service: '염펌',
-      employee: '박미글',
+      staff: '박미글',
       date: '2025.06.08 14시',
       change: '삭제',
       changedAt: '2025.06.08 11:08',
